requirements.txt
pandas~=2.2.2
matplotlib~=3.9.0
seaborn~=0.13.2
openpyxl
numpy~=1.26.4
scikit-learn~=1.5.1
python-dotenv~=1.0.1
PyYAML~=6.0.1
imblearn~=0.0
torch~=2.3.1
fastapi~=0.115.6
uvicorn~=0.32.1
nltk~=3.9.1
langdetect~=1.0.9
imbalanced-learn~=0.12.4
rake_nltk
aiml~=0.9.2
transformers~=4.47.1
starlette~=0.41.3